requirements.txt
mkdocs
pygithub>=1.59.0
parse
